refactor(#8881): split monolithic middleware.js into modular testable components
